+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.25.2] - 2019-10-18
 
 ### Fixed
 - The model's terminating (`$this->terminate();`) feature which was broken because of the forced 404 feature.